SAUCE: HACK: move omap spl base address
Move NON_SECURE_SRAM_START from 0x40304000 to 0x40303000 and
move u-boot-spl load address from 0x40304350 to 0x40303080.

This allows it to stay under 0x4030c000 which is what the TRM
says it must do.  SRAM actually starts at 0x40300000 on GP
devices but was set to 0x40304000 to be the same as Secure
devices so this breaks that.

Signed-off-by: John Rigby <john.rigby@linaro.org>